This is the new state of this app.
GUI is almost finished. As for now, it will do. The "backend" isn't fully implemented, need a lot of work. Thinking to drop wxWidgets, switch to Qt(?). As for "backend", would be better to integrate it with the "frontend" as wxWidgets (also Qt) has it's functions for websockets and is easier to implement cross-platform. (As for Windows and Linux, C++ has other libraries for it, it can be done, but require more work).